[NO-TASK] Bank TX refactoring 2
