+++ src/subdomains/core/sell-crypto/sell-crypto.module.ts
@@ -7,6 +7,7 @@ import { UserModule } from 'src/subdomains/generic/user/user.module';
 import { AddressPoolModule } from 'src/subdomains/supporting/address-pool/address-pool.module';
 import { BankTxModule } from 'src/subdomains/supporting/bank-tx/bank-tx.module';
 import { BankModule } from 'src/subdomains/supporting/bank/bank.module';
+import { FiatOutputModule } from 'src/subdomains/supporting/fiat-output/fiat-output.module';
 import { NotificationModule } from 'src/subdomains/supporting/notification/notification.module';
 import { PayInModule } from 'src/subdomains/supporting/payin/payin.module';
 import { BuyCryptoModule } from '../buy-crypto/buy-crypto.module';
@@ -34,6 +35,7 @@ import { SellService } from './route/sell.service';
     forwardRef(() => PayInModule),
     forwardRef(() => BuyCryptoModule),
     forwardRef(() => AddressPoolModule),
+    FiatOutputModule,
   ],
   controllers: [BuyFiatController, SellController],
   providers: [
