+++ src/subdomains/generic/gs/gs.module.ts
@@ -8,6 +8,7 @@ import { SellCryptoModule } from 'src/subdomains/core/sell-crypto/sell-crypto.mo
 import { AddressPoolModule } from 'src/subdomains/supporting/address-pool/address-pool.module';
 import { BankTxModule } from 'src/subdomains/supporting/bank-tx/bank-tx.module';
 import { BankModule } from 'src/subdomains/supporting/bank/bank.module';
+import { FiatOutputModule } from 'src/subdomains/supporting/fiat-output/fiat-output.module';
 import { NotificationModule } from 'src/subdomains/supporting/notification/notification.module';
 import { PayInModule } from 'src/subdomains/supporting/payin/payin.module';
 import { UserModule } from '../user/user.module';
@@ -30,6 +31,7 @@ import { GsService } from './gs.service';
     BankModule,
     BankTxModule,
     PayInModule,
+    FiatOutputModule,
   ],
   controllers: [GsController, GsEvmController],
   providers: [GsService, GsEvmService],
