+++ src/subdomains/supporting/fiat-output/__mocks__/fiat-output.entity.mock.ts
