+++ src/subdomains/supporting/supporting.module.ts
@@ -3,6 +3,7 @@ import { AddressPoolModule } from './address-pool/address-pool.module';
 import { BankTxModule } from './bank-tx/bank-tx.module';
 import { BankModule } from './bank/bank.module';
 import { DexModule } from './dex/dex.module';
+import { FiatOutputModule } from './fiat-output/fiat-output.module';
 import { FiatPayInModule } from './fiat-payin/fiat-payin.module';
 import { LogModule } from './log/log.module';
 import { NotificationModule } from './notification/notification.module';
@@ -22,6 +23,7 @@ import { PricingModule } from './pricing/pricing.module';
     PayoutModule,
     PricingModule,
     FiatPayInModule,
+    FiatOutputModule,
   ],
   controllers: [],
   providers: [],
